Incorporate notification
