+++ Cargo.toml
@@ -43,6 +43,7 @@ more-asserts = "0.2.2"
 async-trait = "0.1.67"
 tokio-context = "0.1.3"
 mio = "0.8.6"
+parse-display-derive = "0.8.0"
 
 [dev-dependencies]
 rand_07 = { package = "rand", version = "0.7" }
