+++ src/rpc/mod.rs
@@ -1,0 +1,468 @@
+use enr::{CombinedKey, Enr};
+use parse_display_derive::Display;
+use rlp::DecoderError;
+use std::net::{IpAddr, Ipv6Addr};
+use tracing::{debug, warn};
+
+mod notification;
+mod request;
+mod response;
+
+pub use notification::Notification;
+pub use request::{Request, RequestBody, RequestId};
+pub use response::{Response, ResponseBody};
+
+/// Ping notification type.
+pub const PING_MSG_TYPE: u8 = 1;
+/// Pong notification type.
+pub const PONG_MSG_TYPE: u8 = 2;
+/// FindNode notification type.
+pub const FINDNODE_MSG_TYPE: u8 = 3;
+/// Nodes notification type.
+pub const NODES_MSG_TYPE: u8 = 4;
+/// TalkReq notification type.
+pub const TALKREQ_MSG_TYPE: u8 = 5;
+/// TalkResp notification type.
+pub const TALKRESP_MSG_TYPE: u8 = 6;
+/// RelayInit notification type.
+pub const REALYINIT_MSG_TYPE: u8 = 7;
+/// RelayMsg notification type.
+pub const REALYMSG_MSG_TYPE: u8 = 8;
+
+#[derive(Debug, Clone, PartialEq, Eq, Display)]
+/// A combined type representing requests and responses.
+pub enum Message {
+    /// A request, which contains its [`RequestId`].
+    #[display("{0}")]
+    Request(Request),
+    /// A Response, which contains the [`RequestId`] of its associated request.
+    #[display("{0}")]
+    Response(Response),
+    /// A unicast notification.
+    #[display("{0}")]
+    Notification(Notification),
+}
+
+#[allow(dead_code)]
+impl Message {
+    pub fn encode(self) -> Vec<u8> {
+        match self {
+            Self::Request(request) => request.encode(),
+            Self::Response(response) => response.encode(),
+            Self::Notification(notif) => notif.encode(),
+        }
+    }
+
+    pub fn decode(data: &[u8]) -> Result<Self, DecoderError> {
+        if data.len() < 3 {
+            return Err(DecoderError::RlpIsTooShort);
+        }
+
+        let msg_type = data[0];
+
+        let rlp = rlp::Rlp::new(&data[1..]);
+
+        let list_len = rlp.item_count().and_then(|size| {
+            if size < 2 {
+                Err(DecoderError::RlpIncorrectListLen)
+            } else {
+                Ok(size)
+            }
+        })?;
+
+        let id = RequestId::decode(rlp.val_at::<Vec<u8>>(0)?)?;
+
+        let message = match msg_type {
+            1 => {
+                // PingRequest
+                if list_len != 2 {
+                    debug!(
+                        "Ping Request has an invalid RLP list length. Expected 2, found {}",
+                        list_len
+                    );
+                    return Err(DecoderError::RlpIncorrectListLen);
+                }
+                Message::Request(Request {
+                    id,
+                    body: RequestBody::Ping {
+                        enr_seq: rlp.val_at::<u64>(1)?,
+                    },
+                })
+            }
+            2 => {
+                // PingResponse
+                if list_len != 4 {
+                    debug!(
+                        "Ping Response has an invalid RLP list length. Expected 4, found {}",
+                        list_len
+                    );
+                    return Err(DecoderError::RlpIncorrectListLen);
+                }
+                let ip_bytes = rlp.val_at::<Vec<u8>>(2)?;
+                let ip = match ip_bytes.len() {
+                    4 => {
+                        let mut ip = [0u8; 4];
+                        ip.copy_from_slice(&ip_bytes);
+                        IpAddr::from(ip)
+                    }
+                    16 => {
+                        let mut ip = [0u8; 16];
+                        ip.copy_from_slice(&ip_bytes);
+                        let ipv6 = Ipv6Addr::from(ip);
+                        // If the ipv6 is ipv4 compatible/mapped, simply return the ipv4.
+                        if let Some(ipv4) = ipv6.to_ipv4() {
+                            IpAddr::V4(ipv4)
+                        } else {
+                            IpAddr::V6(ipv6)
+                        }
+                    }
+                    _ => {
+                        debug!("Ping Response has incorrect byte length for IP");
+                        return Err(DecoderError::RlpIncorrectListLen);
+                    }
+                };
+                let port = rlp.val_at::<u16>(3)?;
+                Message::Response(Response {
+                    id,
+                    body: ResponseBody::Pong {
+                        enr_seq: rlp.val_at::<u64>(1)?,
+                        ip,
+                        port,
+                    },
+                })
+            }
+            3 => {
+                // FindNodeRequest
+                if list_len != 2 {
+                    debug!(
+                        "FindNode Request has an invalid RLP list length. Expected 2, found {}",
+                        list_len
+                    );
+                    return Err(DecoderError::RlpIncorrectListLen);
+                }
+                let distances = rlp.list_at::<u64>(1)?;
+
+                for distance in distances.iter() {
+                    if distance > &256u64 {
+                        warn!(
+                            "Rejected FindNode request asking for unknown distance {}, maximum 256",
+                            distance
+                        );
+                        return Err(DecoderError::Custom("FINDNODE request distance invalid"));
+                    }
+                }
+
+                Message::Request(Request {
+                    id,
+                    body: RequestBody::FindNode { distances },
+                })
+            }
+            4 => {
+                // NodesResponse
+                if list_len != 3 {
+                    debug!(
+                        "Nodes Response has an invalid RLP list length. Expected 3, found {}",
+                        list_len
+                    );
+                    return Err(DecoderError::RlpIncorrectListLen);
+                }
+
+                let nodes = {
+                    let enr_list_rlp = rlp.at(2)?;
+                    if enr_list_rlp.is_empty() {
+                        // no records
+                        vec![]
+                    } else {
+                        enr_list_rlp.as_list::<Enr<CombinedKey>>()?
+                    }
+                };
+                Message::Response(Response {
+                    id,
+                    body: ResponseBody::Nodes {
+                        total: rlp.val_at::<u64>(1)?,
+                        nodes,
+                    },
+                })
+            }
+            5 => {
+                // Talk Request
+                if list_len != 3 {
+                    debug!(
+                        "Talk Request has an invalid RLP list length. Expected 3, found {}",
+                        list_len
+                    );
+                    return Err(DecoderError::RlpIncorrectListLen);
+                }
+                let protocol = rlp.val_at::<Vec<u8>>(1)?;
+                let request = rlp.val_at::<Vec<u8>>(2)?;
+                Message::Request(Request {
+                    id,
+                    body: RequestBody::TalkReq { protocol, request },
+                })
+            }
+            6 => {
+                // Talk Response
+                if list_len != 2 {
+                    debug!(
+                        "Talk Response has an invalid RLP list length. Expected 2, found {}",
+                        list_len
+                    );
+                    return Err(DecoderError::RlpIncorrectListLen);
+                }
+                let response = rlp.val_at::<Vec<u8>>(1)?;
+                Message::Response(Response {
+                    id,
+                    body: ResponseBody::TalkResp { response },
+                })
+            }
+            _ => {
+                return Err(DecoderError::Custom("Unknown RPC message type"));
+            }
+        };
+
+        Ok(message)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use enr::EnrBuilder;
+
+    #[test]
+    fn ref_test_encode_request_ping() {
+        // reference input
+        let id = RequestId(vec![1]);
+        let enr_seq = 1;
+        let message = Message::Request(Request {
+            id,
+            body: RequestBody::Ping { enr_seq },
+        });
+
+        // expected hex output
+        let expected_output = hex::decode("01c20101").unwrap();
+
+        dbg!(hex::encode(message.clone().encode()));
+        assert_eq!(message.encode(), expected_output);
+    }
+
+    #[test]
+    fn ref_test_encode_request_findnode() {
+        // reference input
+        let id = RequestId(vec![1]);
+        let distances = vec![256];
+        let message = Message::Request(Request {
+            id,
+            body: RequestBody::FindNode { distances },
+        });
+
+        // expected hex output
+        let expected_output = hex::decode("03c501c3820100").unwrap();
+        dbg!(hex::encode(message.clone().encode()));
+
+        assert_eq!(message.encode(), expected_output);
+    }
+
+    #[test]
+    fn ref_test_encode_response_ping() {
+        // reference input
+        let id = RequestId(vec![1]);
+        let enr_seq = 1;
+        let ip: IpAddr = "127.0.0.1".parse().unwrap();
+        let port = 5000;
+        let message = Message::Response(Response {
+            id,
+            body: ResponseBody::Pong { enr_seq, ip, port },
+        });
+
+        // expected hex output
+        let expected_output = hex::decode("02ca0101847f000001821388").unwrap();
+
+        dbg!(hex::encode(message.clone().encode()));
+        assert_eq!(message.encode(), expected_output);
+    }
+
+    #[test]
+    fn ref_test_encode_response_nodes_empty() {
+        // reference input
+        let id = RequestId(vec![1]);
+        let total = 1;
+
+        // expected hex output
+        let expected_output = hex::decode("04c30101c0").unwrap();
+
+        let message = Message::Response(Response {
+            id,
+            body: ResponseBody::Nodes {
+                total,
+                nodes: vec![],
+            },
+        });
+        assert_eq!(message.encode(), expected_output);
+    }
+
+    #[test]
+    fn ref_test_encode_response_nodes() {
+        // reference input
+        let id = RequestId(vec![1]);
+        let total = 1;
+
+        let enr = "-HW4QCjfjuCfSmIJHxqLYfGKrSz-Pq3G81DVJwd_muvFYJiIOkf0bGtJu7kZVCOPnhSTMneyvR4MRbF3G5TNB4wy2ssBgmlkgnY0iXNlY3AyNTZrMaEDymNMrg1JrLQB2KTGtv6MVbcNEVv0AHacwUAPMljNMTg".parse::<Enr<CombinedKey>>().unwrap();
+        // expected hex output
+        let expected_output = hex::decode("04f87b0101f877f875b84028df8ee09f4a62091f1a8b61f18aad2cfe3eadc6f350d527077f9aebc56098883a47f46c6b49bbb91954238f9e14933277b2bd1e0c45b1771b94cd078c32dacb0182696482763489736563703235366b31a103ca634cae0d49acb401d8a4c6b6fe8c55b70d115bf400769cc1400f3258cd3138").unwrap();
+
+        let message = Message::Response(Response {
+            id,
+            body: ResponseBody::Nodes {
+                total,
+                nodes: vec![enr],
+            },
+        });
+        dbg!(hex::encode(message.clone().encode()));
+        assert_eq!(message.encode(), expected_output);
+    }
+
+    #[test]
+    fn ref_test_encode_response_nodes_multiple() {
+        // reference input
+        let id = RequestId(vec![1]);
+        let total = 1;
+        let enr = "enr:-HW4QBzimRxkmT18hMKaAL3IcZF1UcfTMPyi3Q1pxwZZbcZVRI8DC5infUAB_UauARLOJtYTxaagKoGmIjzQxO2qUygBgmlkgnY0iXNlY3AyNTZrMaEDymNMrg1JrLQB2KTGtv6MVbcNEVv0AHacwUAPMljNMTg".parse::<Enr<CombinedKey>>().unwrap();
+
+        let enr2 = "enr:-HW4QNfxw543Ypf4HXKXdYxkyzfcxcO-6p9X986WldfVpnVTQX1xlTnWrktEWUbeTZnmgOuAY_KUhbVV1Ft98WoYUBMBgmlkgnY0iXNlY3AyNTZrMaEDDiy3QkHAxPyOgWbxp5oF1bDdlYE6dLCUUp8xfVw50jU".parse::<Enr<CombinedKey>>().unwrap();
+
+        // expected hex output
+        let expected_output = hex::decode("04f8f20101f8eef875b8401ce2991c64993d7c84c29a00bdc871917551c7d330fca2dd0d69c706596dc655448f030b98a77d4001fd46ae0112ce26d613c5a6a02a81a6223cd0c4edaa53280182696482763489736563703235366b31a103ca634cae0d49acb401d8a4c6b6fe8c55b70d115bf400769cc1400f3258cd3138f875b840d7f1c39e376297f81d7297758c64cb37dcc5c3beea9f57f7ce9695d7d5a67553417d719539d6ae4b445946de4d99e680eb8063f29485b555d45b7df16a1850130182696482763489736563703235366b31a1030e2cb74241c0c4fc8e8166f1a79a05d5b0dd95813a74b094529f317d5c39d235").unwrap();
+
+        let message = Message::Response(Response {
+            id,
+            body: ResponseBody::Nodes {
+                total,
+                nodes: vec![enr, enr2],
+            },
+        });
+        dbg!(hex::encode(message.clone().encode()));
+        assert_eq!(message.encode(), expected_output);
+    }
+
+    #[test]
+    fn ref_decode_response_nodes_multiple() {
+        let input = hex::decode("04f8f20101f8eef875b8401ce2991c64993d7c84c29a00bdc871917551c7d330fca2dd0d69c706596dc655448f030b98a77d4001fd46ae0112ce26d613c5a6a02a81a6223cd0c4edaa53280182696482763489736563703235366b31a103ca634cae0d49acb401d8a4c6b6fe8c55b70d115bf400769cc1400f3258cd3138f875b840d7f1c39e376297f81d7297758c64cb37dcc5c3beea9f57f7ce9695d7d5a67553417d719539d6ae4b445946de4d99e680eb8063f29485b555d45b7df16a1850130182696482763489736563703235366b31a1030e2cb74241c0c4fc8e8166f1a79a05d5b0dd95813a74b094529f317d5c39d235").unwrap();
+
+        let expected_enr1 = "enr:-HW4QBzimRxkmT18hMKaAL3IcZF1UcfTMPyi3Q1pxwZZbcZVRI8DC5infUAB_UauARLOJtYTxaagKoGmIjzQxO2qUygBgmlkgnY0iXNlY3AyNTZrMaEDymNMrg1JrLQB2KTGtv6MVbcNEVv0AHacwUAPMljNMTg".parse::<Enr<CombinedKey>>().unwrap();
+        let expected_enr2 = "enr:-HW4QNfxw543Ypf4HXKXdYxkyzfcxcO-6p9X986WldfVpnVTQX1xlTnWrktEWUbeTZnmgOuAY_KUhbVV1Ft98WoYUBMBgmlkgnY0iXNlY3AyNTZrMaEDDiy3QkHAxPyOgWbxp5oF1bDdlYE6dLCUUp8xfVw50jU".parse::<Enr<CombinedKey>>().unwrap();
+
+        let decoded = Message::decode(&input).unwrap();
+
+        match decoded {
+            Message::Response(response) => match response.body {
+                ResponseBody::Nodes { total, nodes } => {
+                    assert_eq!(total, 1);
+                    assert_eq!(nodes[0], expected_enr1);
+                    assert_eq!(nodes[1], expected_enr2);
+                }
+                _ => panic!("Invalid decoding"),
+            },
+            _ => panic!("Invalid decoding"),
+        }
+    }
+
+    #[test]
+    fn encode_decode_ping_request() {
+        let id = RequestId(vec![1]);
+        let request = Message::Request(Request {
+            id,
+            body: RequestBody::Ping { enr_seq: 15 },
+        });
+
+        let encoded = request.clone().encode();
+        let decoded = Message::decode(&encoded).unwrap();
+
+        assert_eq!(request, decoded);
+    }
+
+    #[test]
+    fn encode_decode_ping_response() {
+        let id = RequestId(vec![1]);
+        let request = Message::Response(Response {
+            id,
+            body: ResponseBody::Pong {
+                enr_seq: 15,
+                ip: "127.0.0.1".parse().unwrap(),
+                port: 80,
+            },
+        });
+
+        let encoded = request.clone().encode();
+        let decoded = Message::decode(&encoded).unwrap();
+
+        assert_eq!(request, decoded);
+    }
+
+    #[test]
+    fn encode_decode_find_node_request() {
+        let id = RequestId(vec![1]);
+        let request = Message::Request(Request {
+            id,
+            body: RequestBody::FindNode {
+                distances: vec![12],
+            },
+        });
+
+        let encoded = request.clone().encode();
+        let decoded = Message::decode(&encoded).unwrap();
+
+        assert_eq!(request, decoded);
+    }
+
+    #[test]
+    fn encode_decode_nodes_response() {
+        let key = CombinedKey::generate_secp256k1();
+        let enr1 = EnrBuilder::new("v4")
+            .ip4("127.0.0.1".parse().unwrap())
+            .udp4(500)
+            .build(&key)
+            .unwrap();
+        let enr2 = EnrBuilder::new("v4")
+            .ip4("10.0.0.1".parse().unwrap())
+            .tcp4(8080)
+            .build(&key)
+            .unwrap();
+        let enr3 = EnrBuilder::new("v4")
+            .ip("10.4.5.6".parse().unwrap())
+            .build(&key)
+            .unwrap();
+
+        let enr_list = vec![enr1, enr2, enr3];
+        let id = RequestId(vec![1]);
+        let request = Message::Response(Response {
+            id,
+            body: ResponseBody::Nodes {
+                total: 1,
+                nodes: enr_list,
+            },
+        });
+
+        let encoded = request.clone().encode();
+        let decoded = Message::decode(&encoded).unwrap();
+
+        assert_eq!(request, decoded);
+    }
+
+    #[test]
+    fn encode_decode_talk_request() {
+        let id = RequestId(vec![1]);
+        let request = Message::Request(Request {
+            id,
+            body: RequestBody::TalkReq {
+                protocol: vec![17u8; 32],
+                request: vec![1, 2, 3],
+            },
+        });
+
+        let encoded = request.clone().encode();
+        let decoded = Message::decode(&encoded).unwrap();
+
+        assert_eq!(request, decoded);
+    }
+}
