+++ src/rpc/request/request_body.rs
@@ -1,0 +1,37 @@
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum RequestBody {
+    /// A PING request.
+    Ping {
+        /// Our current ENR sequence number.
+        enr_seq: u64,
+    },
+    /// A FINDNODE request.
+    FindNode {
+        /// The distance(s) of peers we expect to be returned in the response.
+        distances: Vec<u64>,
+    },
+    /// A Talk request.
+    TalkReq {
+        /// The protocol requesting.
+        protocol: Vec<u8>,
+        /// The request.
+        request: Vec<u8>,
+    },
+}
+
+impl std::fmt::Display for RequestBody {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            RequestBody::Ping { enr_seq } => write!(f, "PING: enr_seq: {enr_seq}"),
+            RequestBody::FindNode { distances } => {
+                write!(f, "FINDNODE Request: distance: {distances:?}")
+            }
+            RequestBody::TalkReq { protocol, request } => write!(
+                f,
+                "TALK: protocol: {}, request: {}",
+                hex::encode(protocol),
+                hex::encode(request)
+            ),
+        }
+    }
+}
