+++ src/rpc/response/response_body.rs
@@ -1,0 +1,54 @@
+use crate::Enr;
+use std::net::IpAddr;
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum ResponseBody {
+    /// A PONG response.
+    Pong {
+        /// The current ENR sequence number of the responder.
+        enr_seq: u64,
+        /// Our external IP address as observed by the responder.
+        ip: IpAddr,
+        /// Our external UDP port as observed by the responder.
+        port: u16,
+    },
+    /// A NODES response.
+    Nodes {
+        /// The total number of responses that make up this response.
+        total: u64,
+        /// A list of ENR's returned by the responder.
+        nodes: Vec<Enr>,
+    },
+    /// The TALK response.
+    TalkResp {
+        /// The response for the talk.
+        response: Vec<u8>,
+    },
+}
+
+impl std::fmt::Display for ResponseBody {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            ResponseBody::Pong { enr_seq, ip, port } => {
+                write!(f, "PONG: Enr-seq: {enr_seq}, Ip: {ip:?},  Port: {port}")
+            }
+            ResponseBody::Nodes { total, nodes } => {
+                write!(f, "NODES: total: {total}, Nodes: [")?;
+                let mut first = true;
+                for id in nodes {
+                    if !first {
+                        write!(f, ", {id}")?;
+                    } else {
+                        write!(f, "{id}")?;
+                    }
+                    first = false;
+                }
+
+                write!(f, "]")
+            }
+            ResponseBody::TalkResp { response } => {
+                write!(f, "Response: Response {}", hex::encode(response))
+            }
+        }
+    }
+}
